+++ Libplanet.Action/AssemblyInfo.cs
@@ -3,3 +3,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Libplanet.Tests")]
 [assembly: InternalsVisibleTo("Libplanet.Action.Tests")]
 [assembly: InternalsVisibleTo("Libplanet.Explorer.Tests")]
+[assembly: InternalsVisibleTo("Libplanet.Mocks")]