+++ Libplanet.Explorer.Tests/Queries/StateQueryTest.cs
@@ -9,6 +9,7 @@ using Libplanet.Action.State;
 using Libplanet.Common;
 using Libplanet.Crypto;
 using Libplanet.Explorer.Queries;
+using Libplanet.Types.Assets;
 using Libplanet.Types.Consensus;
 using Xunit;
 using static Libplanet.Explorer.Tests.GraphQLTestUtils;
@@ -138,12 +139,7 @@ public partial class StateQueryTest
             ByteUtil.Hex(_codec.Encode(new Integer(123))),
             Assert.IsAssignableFrom<string>(
                 Assert.IsAssignableFrom<IDictionary<string, object>>(balances[0])["hex"]));
-
-        // FIXME: Due to dumb mocking. We need to overhaul mocking.
-        Assert.Equal(
-            ByteUtil.Hex(_codec.Encode(new Integer(123))),
-            Assert.IsAssignableFrom<string>(
-                Assert.IsAssignableFrom<IDictionary<string, object>>(balances[1])["hex"]));
+        Assert.Null(balances[1]);
     }
 
     [Fact]