+++ Libplanet.Tests/Action/AccountTest.cs
@@ -1,11 +1,8 @@
-using System.Collections.Immutable;
 using System.Linq;
 using Bencodex.Types;
-using Libplanet.Action;
 using Libplanet.Action.State;
-using Libplanet.Action.Tests.Mocks;
 using Libplanet.Crypto;
-using Libplanet.Types.Assets;
+using Libplanet.Mocks;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -15,10 +12,7 @@ namespace Libplanet.Tests.Action
     {
         private readonly PrivateKey[] _keys;
         private readonly Address[] _addr;
-        private readonly Currency[] _currencies;
         private readonly IAccount _initAccount;
-        private readonly IActionContext _initContext;
-        private readonly Address _accountAddress = ReservedAddresses.LegacyAccount;
 
         public AccountTest(ITestOutputHelper output)
         {
@@ -30,21 +24,9 @@ namespace Libplanet.Tests.Action
             };
 
             _addr = _keys.Select(key => key.Address).ToArray();
-
-            _currencies = new[]
-            {
-#pragma warning disable CS0618  // must test obsoleted Currency.Legacy() for backwards compatibility
-                Currency.Legacy("FOO", 0, _addr[0]),
-                Currency.Legacy("BAR", 0, _addr.Take(2).ToImmutableHashSet()),
-                Currency.Legacy("BAZ", 0, null),
-#pragma warning restore CS0618  // must test obsoleted Currency.Legacy() for backwards compatibility
-                Currency.Uncapped("QUX", 0, minter: _addr[0]),
-                Currency.Capped("QUUX", 0, (100, 0), minter: _addr[0]),
-            };
-
-            _initAccount = new Account(new MockAccountState()
+            _initAccount = new Account(MockUtil.MockAccountState)
                 .SetState(_addr[0], (Text)"a")
-                .SetState(_addr[1], (Text)"b"));
+                .SetState(_addr[1], (Text)"b");
 
             output.WriteLine("Fixtures  {0,-42}  State", "Address");
             int i = 0;
@@ -56,23 +38,6 @@ namespace Libplanet.Tests.Action
                     a,
                     _initAccount.GetStates(new[] { a })[0]);
             }
-
-            _initContext = CreateContext(_initAccount, _addr[0]);
-        }
-
-        public IActionContext CreateContext(IAccount delta, Address signer)
-        {
-            IWorld world = new World(new MockWorldState());
-            world = world.SetAccount(_accountAddress, delta);
-            return new ActionContext(
-                signer,
-                null,
-                signer,
-                0,
-                0,
-                world,
-                0,
-                0);
         }
 
         [Fact]